% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decode.R
\name{decode}
\alias{decode}
\title{Decode Polyline}
\usage{
decode(polylines, precision = 5)
}
\arguments{
\item{polylines}{vector of encoded polyline strings}

\item{precision}{number of decimal places the coordinates were encoded
with (5 is the precision used by the Google polyline encoding algorithm)}
}
\description{
Decodes encoded polylines into a list of data.frames.
}
\examples{
polylines <- c(
  "ohlbDnbmhN~suq@am{tAw`qsAeyhGvkz`@fge}A",
  "ggmnDt}wmLgc`DesuQvvrLofdDorqGtzzV"
)

decode(polylines)

}
