% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decode.R
\name{decodePolyline}
\alias{decodePolyline}
\title{Decode Polyline}
\usage{
decodePolyline(polylines)
}
\arguments{
\item{polylines}{}
}
\description{
Decodes encoded polylines
}
\examples{
polylines <- c(
  "ohlbDnbmhN~suq@am{tAw`qsAeyhGvkz`@fge}A",
  "ggmnDt}wmLgc`DesuQvvrLofdDorqGtzzV"
)

decodePolyline(polylines)

}
