% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeSf.R
\name{encodeSf}
\alias{encodeSf}
\title{Encode sf}
\usage{
encodeSf(sf)
}
\arguments{
\item{sf}{object}
}
\value{
\code{sfencoded} object
}
\description{
Converts an sf object into an encoded sf object
}
\examples{

}
