% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{polyline_wkt}
\alias{polyline_wkt}
\title{Polyline WKT}
\usage{
polyline_wkt(obj)
}
\arguments{
\item{obj}{\code{sfencoded} object or \code{encoded_column} of encoded polylines}
}
\value{
well-known text representation of the encoded polylines
}
\description{
Converts encoded polylines into well-known text. 
Note that 'polylines' refers to encoded strings here.
Valid inputs include any `sf` object with an encoded column.
}
\details{
The function also assumes Google Web Mercator projection (WSG 84 / EPSG:3857 / EPSG:900913)
for inputs and outputs.
}
