% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_sf.R
\name{sfAttributes}
\alias{sfAttributes}
\title{sf Attributes}
\usage{
sfAttributes(x)
}
\arguments{
\item{x}{\code{sfencoded} object}
}
\value{
\code{list} of \code{sf} attributes
}
\description{
Retrieves the sf attributes stored on the \code{sfencoded} object
}
