% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{wkt_polyline}
\alias{wkt_polyline}
\title{WKT Polyline}
\usage{
wkt_polyline(obj)
}
\arguments{
\item{obj}{\code{sfencoded} object or \code{wkt_column} of well-known text}
}
\value{
encoded polyline representation of geometries
}
\description{
WKT Polyline
}
