% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decode.R
\name{decode}
\alias{decode}
\title{Decode Polyline}
\usage{
decode(polylines)
}
\arguments{
\item{polylines}{vector of encoded polyline strings}
}
\description{
Decodes encoded polylines into a list of data.frames.
}
\examples{
polylines <- c(
  "ohlbDnbmhN~suq@am{tAw`qsAeyhGvkz`@fge}A",
  "ggmnDt}wmLgc`DesuQvvrLofdDorqGtzzV"
)

decode(polylines)

}
