% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encode.R
\name{encode}
\alias{encode}
\title{Encode}
\usage{
encode(obj, strip = NA, lon = NA, lat = NA)
}
\arguments{
\item{obj}{object}

\item{strip}{(optional) logical indicating if \code{sf} attributes should be stripped. 
Useful if the object contains only one type of geometry and you want to reduce the size
even further}

\item{lon}{(optional) vector of longitudes}

\item{lat}{(optional) vector of latitudes}
}
\value{
encoded object
}
\description{
Encodes coordinates into an encoded polyline
}
\examples{
\dontrun{

## sf objects
library(sf)
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))

encoded <- encode(nc)

## strip off attributes
encodedLite <- encode(nc, TRUE)


}


}
