% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encode.R
\name{encode}
\alias{encode}
\title{Encode sf}
\usage{
encode(sf, strip)
}
\arguments{
\item{sf}{object}

\item{strip}{logical indicating if attributes should be stripped. Useful if
the object contains only one type of geometry and you want to reduce the size
even further}
}
\value{
\code{sfencoded} object
}
\description{
Converts an sf object into an encoded sf object
}
\examples{

}
