% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{polyline_wkt}
\alias{polyline_wkt}
\title{Polyline WKT}
\usage{
polyline_wkt(obj)
}
\arguments{
\item{obj}{\code{sfencoded} object or \code{encoded_column} of encoded polylines}
}
\value{
well-known text representation of the encoded polylines
}
\description{
Converts encoded polylines into well-known text
}
